Add files via upload
Good evening ma'am. This is my prelim exam hands on, which is the second part that is all about java, and I was not able to finish during the exam and was only able to accomplish half of it. Yun lang po, thank you po. Hope you have a nice day.